+++ cloud/filestore/tests/async_open_noshards_test/test.py
@@ -1,7 +1,5 @@
-import runpy
-
-import yatest.common as common
+from cloud.filestore.tests.async_open_test.lib import async_test
 
 
-test = runpy.run_path(common.source_path(
-    "cloud/filestore/tests/async_open_test/test.py"))["test"]
+def test():
+    async_test()
